Detect vsnprintf() support for "%m" and disable our "%m" expansion.
Add --enable-c99-sprintf to configure args for -noopenssl variety of
flock-build to avoid regressions in (v)snprintf() replacement.

More msnprintf() unit tests.

Coverity Scan error checking fixes.

+++ ntpq/ntpq.c
@@ -963,9 +963,9 @@ getresponse(
 		}
 
 		TRACE(2, ("Got packet, size = %d\n", n));
-		if ((int)count > (n - CTL_HEADER_LEN)) {
-			TRACE(1, ("Received count of %u octets, data in packet is %d\n",
-				  count, (n - CTL_HEADER_LEN)));
+		if ((long)count > (n - CTL_HEADER_LEN)) {
+			TRACE(1, ("Received count of %u octets, data in packet is %ld\n",
+				  count, (long)n - CTL_HEADER_LEN));
 			continue;
 		}
 		if (count == 0 && CTL_ISMORE(rpkt.r_m_e_op)) {
